worker interval changed, new tests added
